+++ src/collective/recaptcha/__init__.py
@@ -1,5 +1,8 @@
 # coding=utf-8
+from Products.CMFPlone.utils import getFSVersionTuple
 from zope.i18nmessageid import MessageFactory
 
 
 RecaptchaMessageFactory = MessageFactory("collective.recaptcha")
+
+PLONE4 = getFSVersionTuple()[0] == 4