+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 2.1.1 (unreleased)
 ------------------
 
+- Fix recaptcha loading in Plone 4 modal.
+  [wesleybl]
+
 - Fixed startup on Python 3 (Plone 5.2).
   The tests are not run yet on 5.2, so compatibility is not confirmed.
   [maurits]